Tokenizer working!
Finally got the interpreter to properly split tokens without
whitespace. This made classifying tokens as operators or operands very
easy. Now I have to figure out how to determine which operators are
unary and which are binary, and how they interact. After that, I still
have to make sure that the interpreter can read numbers and signs like
+-*/, which it still treats as strings.